Retagging to version 1.5 for trunk.
We just noticed the style was set to not have new lines at the end, this has been amended. This is needed because the scripts to convert versions will add blank lines to the end of the file. See ACCUMULO-41. I will not be reformatting all 1.4 branch files to the new standard now, but they may be touched next time this script is run. Please be sure to update your formatter file.

git-svn-id: https://svn.apache.org/repos/asf/incubator/accumulo/trunk@1201953 13f79535-47bb-0310-9956-ffa450edef68